Refactor PostProcessing
