+++ inductiva/fluids/post_processing/__init__.py
@@ -1,0 +1,2 @@
+#pylint: disable=missing-module-docstring
+from .openfoam import SteadyStateOutput
