+++ inductiva/fluids/scenarios/__init__.py
@@ -3,6 +3,7 @@ from .dam_break import DamBreak
 from .fluid_block import FluidBlock
 from .wind_tunnel import WindTunnel
 from .wind_tunnel import WindTunnelOutput
+from .wind_terrain import WindOverTerrain
 from .fluid_tank import (
     FluidTank,
     CubicTankOutlet,
