+++ inductiva/fluids/scenarios/wind_tunnel/__init__.py
@@ -1,3 +1,3 @@
 #pylint: disable=missing-module-docstring
 from .wind_tunnel import WindTunnel
-from .post_processing import WindTunnelOutput
+from .output import WindTunnelOutput
