+++ inductiva/fluids/scenarios/wind_tunnel/wind_tunnel.py
@@ -1,7 +1,7 @@
 """Wind tunnel scenario to run an object over an air flow."""
 
 from dataclasses import dataclass
-from enum import Enum
+import enum
 from functools import singledispatchmethod
 import os
 import shutil
@@ -10,25 +10,18 @@ from typing import Optional, List, Literal
 
 from absl import logging
 
-from inductiva import tasks, resources
-from inductiva.types import Path
-from inductiva.scenarios import Scenario
-from inductiva.simulation import Simulator
-from inductiva.fluids.simulators import OpenFOAM
-from inductiva.utils.templates import (TEMPLATES_PATH,
-                                       batch_replace_params_in_template,
-                                       replace_params_in_template)
-from inductiva.utils import files
-from inductiva.fluids.scenarios.wind_tunnel.post_processing import WindTunnelOutput
+import inductiva
+from inductiva import tasks, resources, fluids, types
+from inductiva.utils import templates, files
 
-SCENARIO_TEMPLATE_DIR = os.path.join(TEMPLATES_PATH, "wind_tunnel")
+SCENARIO_TEMPLATE_DIR = os.path.join(templates.TEMPLATES_PATH, "wind_tunnel")
 OPENFOAM_TEMPLATE_INPUT_DIR = "openfoam"
 FILES_SUBDIR = "files"
 COMMANDS_TEMPLATE_FILE_NAME = "commands.json.jinja"
 
 
 @dataclass
-class MeshResolution(Enum):
+class MeshResolution(enum.Enum):
     """Sets particle radius according to resolution."""
     HIGH = [5, 6]
     MEDIUM = [4, 5]
@@ -36,7 +29,7 @@ class MeshResolution(Enum):
     VERY_LOW = [2, 3]
 
 
-class WindTunnel(Scenario):
+class WindTunnel(inductiva.scenarios.Scenario):
     """Physical scenario of a configurable wind tunnel simulation.
 
     A wind tunnel is a tool used in aerodynamic research to study the
@@ -66,7 +59,7 @@ class WindTunnel(Scenario):
     pressure_field, cutting planes and force coefficients.
     """
 
-    valid_simulators = [OpenFOAM]
+    valid_simulators = [fluids.simulators.OpenFOAM]
 
     def __init__(self,
                  flow_velocity: List[float] = None,
@@ -109,10 +102,11 @@ class WindTunnel(Scenario):
         ]
 
     def simulate(self,
-                 simulator: Simulator = OpenFOAM("windtunnel"),
+                 simulator: inductiva.Simulator = fluids.simulators.OpenFOAM(
+                     "windtunnel"),
                  machine_group: Optional[resources.MachineGroup] = None,
                  run_async: bool = False,
-                 object_path: Optional[Path] = None,
+                 object_path: Optional[types.Path] = None,
                  num_iterations: float = 100,
                  resolution: Literal["high", "medium", "low"] = "medium",
                  n_cores: int = 1) -> tasks.Task:
@@ -148,7 +142,7 @@ class WindTunnel(Scenario):
                                 n_cores=n_cores,
                                 commands=commands)
 
-        task.set_output_class(WindTunnelOutput)
+        task.set_output_class(fluids.scenarios.WindTunnelOutput)
         task.set_default_output_files(self.set_default_output_files())
 
         return task
@@ -161,7 +155,7 @@ class WindTunnel(Scenario):
                                               COMMANDS_TEMPLATE_FILE_NAME)
 
         with tempfile.NamedTemporaryFile() as commands_file:
-            replace_params_in_template(
+            templates.replace_params_in_template(
                 template_path=commands_template_path,
                 params={"n_cores": self.n_cores},
                 output_file_path=commands_file.name,
@@ -172,12 +166,12 @@ class WindTunnel(Scenario):
         return commands
 
     @singledispatchmethod
-    def create_input_files(self, simulator: Simulator):
+    def create_input_files(self, simulator: inductiva.Simulator):
         pass
 
 
 @WindTunnel.create_input_files.register
-def _(self, simulator: OpenFOAM, input_dir):  # pylint: disable=unused-argument
+def _(self, simulator: fluids.simulators.OpenFOAM, input_dir):  # pylint: disable=unused-argument
     """Creates OpenFOAM simulation input files."""
 
     # The WindTunnel with OpenFOAM requires changing multiple files
@@ -190,7 +184,7 @@ def _(self, simulator: OpenFOAM, input_dir):  # pylint: disable=unused-argument
                     dirs_exist_ok=True,
                     symlinks=True)
 
-    batch_replace_params_in_template(
+    templates.batch_replace_params_in_template(
         templates_dir=input_dir,
         template_filenames=[
             os.path.join("0", "include",
